requirements/dev.txt
-r base.txt
pycounters
flask-debugtoolbar
Flask-Testing
